requirements.txt
albucore==0.0.23
albumentations==2.0.5
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
bokeh==3.7.2
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
codecarbon==2.8.3
colorlog==6.9.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
Deprecated==1.2.18
docker-pycreds==0.4.0
exceptiongroup==1.2.2
fief-client==0.20.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
ga4mp==2.0.4
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.4.2
jwcrypto==1.5.6
kiwisolver==1.4.8
lightning-utilities==0.14.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
moviepy==2.1.2
mpmath==1.3.0
narwhals==1.35.0
networkx==3.4.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optuna==4.2.1
optuna-dashboard==0.18.0
optuna-integration==4.2.1
optunahub==0.2.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.7
plotly==6.0.1
proglog==0.1.11
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
RapidFuzz==3.12.2
rdkit==2024.9.6
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.24.1
setproctitle==1.3.5
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
SQLAlchemy==2.0.40
stringzilla==3.12.4
sympy==1.13.1
termcolor==2.3.0
threadpoolctl==3.6.0
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
triton==3.2.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.98
ultralytics-thop==2.0.14
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
wrapt==1.17.2
xyzservices==2025.1.0
yaspin==3.1.0
